requirements.txt
Flask==1.1.2
numpy==1.20.3
tensorflow-cpu==2.8.0
tensorflow_hub==0.12.0
